+++ google/externalaccount/testdata/3pi_cred.json
